Populate compositional pressure and pore-volume summaries
The compositional output module left updateFluidInPlace empty, so FPR,
FPRP, RPR, FHPV, and FRPV were reported as zero.

The pore-volume and pressure accumulation used by the black-oil module is
formulation-independent. Move it to GenericOutputModule and call it from
both output modules. For compositional models, use the current porosity as
the reference porosity because their intensive quantities do not expose a
separate value.

Compositional phase in-place volumes remain unsupported because they
require a flash to surface conditions.
