+++ opm/simulators/flow/OutputBlackoilModule.hpp
@@ -1064,71 +1064,6 @@ private:
         return rates;
     }
 
-    template <typename FluidState>
-    Scalar hydroCarbonFraction(const FluidState& fs) const
-    {
-        if (this->eclState_.runspec().co2Storage()) {
-            // CO2 storage: Hydrocarbon volume is full pore-volume.
-            return 1.0;
-        }
-
-        // Common case.  Hydrocarbon volume is fraction occupied by actual
-        // hydrocarbons.
-        auto hydrocarbon = Scalar {0};
-        if (FluidSystem::phaseIsActive(oilPhaseIdx)) {
-            hydrocarbon += getValue(fs.saturation(oilPhaseIdx));
-        }
-
-        if (FluidSystem::phaseIsActive(gasPhaseIdx)) {
-            hydrocarbon += getValue(fs.saturation(gasPhaseIdx));
-        }
-
-        return hydrocarbon;
-    }
-
-    void updateTotalVolumesAndPressures_(const unsigned             globalDofIdx,
-                                         const IntensiveQuantities& intQuants,
-                                         const double               totVolume)
-    {
-        const auto& fs = intQuants.fluidState();
-
-        const double pv = totVolume * intQuants.porosity().value();
-        const auto hydrocarbon = this->hydroCarbonFraction(fs);
-
-        this->fipC_.assignPoreVolume(globalDofIdx,
-                                totVolume * intQuants.referencePorosity(),
-                                pv);
-        if (! this->hydrocarbonPoreVolume_.empty()) {
-            this->hydrocarbonPoreVolume_[globalDofIdx] = pv * hydrocarbon;
-        }
-
-        if (!this->pressureTimesHydrocarbonVolume_.empty() &&
-            !this->pressureTimesPoreVolume_.empty())
-        {
-            assert(this->hydrocarbonPoreVolume_.size() == this->pressureTimesHydrocarbonVolume_.size());
-            assert(this->fipC_.get(Inplace::Phase::PoreVolume).size() == this->pressureTimesPoreVolume_.size());
-
-            if (FluidSystem::phaseIsActive(oilPhaseIdx)) {
-                this->pressureTimesPoreVolume_[globalDofIdx] =
-                    getValue(fs.pressure(oilPhaseIdx)) * pv;
-
-                this->pressureTimesHydrocarbonVolume_[globalDofIdx] =
-                    this->pressureTimesPoreVolume_[globalDofIdx] * hydrocarbon;
-            }
-            else if (FluidSystem::phaseIsActive(gasPhaseIdx)) {
-                this->pressureTimesPoreVolume_[globalDofIdx] =
-                    getValue(fs.pressure(gasPhaseIdx)) * pv;
-
-                this->pressureTimesHydrocarbonVolume_[globalDofIdx] =
-                    this->pressureTimesPoreVolume_[globalDofIdx] * hydrocarbon;
-            }
-            else if (FluidSystem::phaseIsActive(waterPhaseIdx)) {
-                this->pressureTimesPoreVolume_[globalDofIdx] =
-                    getValue(fs.pressure(waterPhaseIdx)) * pv;
-            }
-        }
-    }
-
     void updatePhaseInplaceVolumes_(const unsigned             globalDofIdx,
                                     const IntensiveQuantities& intQuants,
                                     const double               totVolume)
